Added support for LCD displays using the Hitachi HD44780 controller.
The standard Arduino LiquidCrystal library is used to manage the
LCD display.

An LCD display is not required to be connected to the Teensy, but
if it is present, it will be used to display status information.
With an LCD display it is no longer necessary to monitor the
serial port for progress and the cracked pin, it will be shown
on the display.

Some general code clean up as well to eliminate warnings, no
algorithmic changes.